backend/requirements.txt
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-kms==2.16.2
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
crcmod==1.7
cryptography==41.0.4
defusedxml==0.7.1
diff-match-patch==20230430
Django==3.2.16
django-cors-headers==4.3.1
django-filter==23.5
django-import-export==3.3.1
django-js-asset==2.1.0
django-mptt==0.14.0
django-oss-storage==1.1.1
django-simpleui==2023.8.28
django-taggit==1.3.0
django-taggit-serializer==0.1.7
djangorestframework==3.14.0
et-xmlfile==1.1.0
gunicorn==21.2.0
idna==3.4
jmespath==0.10.0
Markdown==3.5.1
MarkupPy==1.14
mysqlclient==2.2.0
odfpy==1.4.1
openpyxl==3.1.2
oss2==2.18.2
packaging==23.1
Pillow==10.0.1
pycparser==2.21
pycryptodome==3.19.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.8.0
urllib3==2.0.5
uWSGI==2.0.23
xlrd==2.0.1
xlwt==1.3.0
